requirements.txt
Django==5.0.2
django-crispy-forms==2.1
django-filter==23.5
Pillow==10.2.0
django-notifications-hq==1.8.3
reportlab==4.1.0
python-dateutil==2.8.2 